requirements.txt
djangorestframework-simplejwt
asgiref
async-timeout
bcrypt
# black
# boto3
# botocore
cffi
# cryptography==37.0.2  # Let system use default version
certifi
charset-normalizer
click
colorama
commonmark
Django
django-cors-headers
django-rest-knox
django-rest-passwordreset
django-restframework
# decorator==4.4.2
# Deprecated==1.2.13
gevent
greenlet
idna
# imageio==2.22.3
# imageio-ffmpeg==0.4.7
# moviepy==1.0.3
# mutagen==1.46.0
# mypy-extensions==0.4.3
# numpy
phonenumbers
# jmespath==1.0.0
#parallel-ssh==2.10.0
paramiko
psycopg2-binary
pycparser
PyNaCl
python-dateutil
pytz
packaging
# pathspec==0.10.1
# Pillow==9.3.0
# platformdirs==2.5.2
# proglog==0.1.10
# Pygments==2.13.0
# pyparsing==3.0.9
# pytube
redis
requests
# rich==12.6.0
# s3transfer==0.6.0
six
sqlparse
# spotipy==2.21.0
urllib3
zope.event
zope.interface
gunicorn
tzdata
django-jsonfield
knox
Jinja2
drf-yasg
djangorestframework
channels
websockets
psutil
channels-redis
# social-auth-app-django  # Comment out to avoid dependency conflicts
# openai
# ray[data,train,tune,serve,default]
# selenium==4.2.0
# xvfbwrapper
# webdriver-manager
# transformers
# sentencepiece
# datasets
# soundfile
# pydub
# pyttsx3
# whisper
# whisper_mic
# torch
# sounddevice
daphne
# playsound
# vosk
# pyautogui
stripe
django-phonenumber-field
Twisted[tls,http2]
# pretty
arrow
tqdm
# wrapt==1.14.1  # Comment out to avoid dependency conflicts
# django-rest-auth
